fix type field

+++ admin/telegram_users/models/history_words.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from django.db.models import BooleanField, CASCADE, DateTimeField, ForeignKey, Model, PositiveIntegerField
+from django.db.models import BooleanField, CASCADE, DateTimeField, FloatField, ForeignKey, Model, PositiveIntegerField
 
 from telegram_users.models.telegram_users import TelegramUsersModel
 from books.models.words import WordsModel
@@ -16,7 +16,7 @@ class UsersWordsHistoryModel(Model):
     correct_answers = PositiveIntegerField(default=0)
     incorrect_answers = PositiveIntegerField(default=0)
     correct_answers_in_row = PositiveIntegerField(default=0)
-    increase_factor = PositiveIntegerField(default=2)
+    increase_factor = FloatField(default=2.0)
     interval_repeat = PositiveIntegerField(default=600)
     repeat_datetime = DateTimeField(default=datetime.utcnow)
     created_at = DateTimeField(default=datetime.utcnow)
